(#777) Moved LoggingUtils from util to transport/spi (so same place as LoggingFilter), putting back the original version in gateway.util to avoid breaking external consumers but marking it as deprecated. Moved the static log methods out of Utils back into LoggingFilter since they are not used elsewhere and are not a good pattern to follow generally. Moved most of the static methods out of LoggingFilter into LoggingUtils so they can be used anywhere to include the session details in any log message.
