+++ transport/spi/src/test/java/org/kaazing/gateway/transport/LoggingFilterTest.java
@@ -16,16 +16,13 @@
 package org.kaazing.gateway.transport;
 
 import static org.junit.Assert.assertEquals;
-import static org.kaazing.gateway.resource.address.ResourceAddress.IDENTITY_RESOLVER;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 
 import java.io.IOException;
 import java.net.Inet6Address;
-import java.net.InetAddress;
 import java.net.InetSocketAddress;
-import java.net.SocketAddress;
-import java.security.Principal;
-import java.util.ArrayList;
-import java.util.Collection;
+import java.util.concurrent.atomic.AtomicReference;
 
 import javax.security.auth.Subject;
 
@@ -33,20 +30,13 @@ import org.apache.mina.core.filterchain.IoFilter.NextFilter;
 import org.apache.mina.core.filterchain.IoFilterChain;
 import org.apache.mina.core.session.IoSession;
 import org.jmock.integration.junit4.JUnitRuleMockery;
-import org.jmock.lib.legacy.ClassImposteriser;
 import org.junit.Rule;
 import org.junit.Test;
-import org.kaazing.gateway.resource.address.IdentityResolver;
 import org.kaazing.gateway.resource.address.ResourceAddress;
-import org.kaazing.gateway.resource.address.ResourceAddressFactories;
 import org.kaazing.gateway.resource.address.ResourceAddressFactory;
-import org.kaazing.gateway.resource.address.ResourceOptions;
-import org.kaazing.gateway.resource.address.http.HttpIdentityResolver;
-import org.kaazing.gateway.security.auth.config.parse.DefaultUserConfig;
 import org.kaazing.gateway.transport.test.Expectations;
 import org.kaazing.mina.core.service.IoAcceptorEx;
 import org.kaazing.mina.core.service.IoConnectorEx;
-import org.kaazing.mina.core.service.IoServiceEx;
 import org.kaazing.mina.core.session.IoSessionEx;
 import org.slf4j.Logger;
 
@@ -62,6 +52,75 @@ public class LoggingFilterTest {
     final Logger logger = context.mock(Logger.class);
 
     @Test
+    public void addIfNeeded_shouldNotAddIfInfoNotEnabled() throws Exception {
+        context.checking(new Expectations() {
+            {
+                oneOf(logger).isInfoEnabled(); will(returnValue(false));
+            }
+        });
+
+        assertFalse(LoggingFilter.addIfNeeded(logger, session, "wsn"));
+    }
+
+    @Test
+    public void addIfNeeded_shouldAddExceptionLoggingFilter() throws Exception {
+        final IoFilterChain filterChain = context.mock(IoFilterChain.class, "filterChain");
+        final IoAcceptorEx service = context.mock(IoAcceptorEx.class, "service");
+        ResourceAddressFactory addressFactory = ResourceAddressFactory.newResourceAddressFactory();
+        final ResourceAddress localAddress = addressFactory.newResourceAddress("http://111.122.133.144:2121/jms");
+        final ResourceAddress  remoteAddress = addressFactory.newResourceAddress("http://111.122.133.144:41234/jms");
+        final Subject subject = new Subject();
+
+        AtomicReference<LoggingFilter> filter = new AtomicReference<LoggingFilter>();
+
+        context.checking(new Expectations() {
+            {
+                oneOf(logger).isInfoEnabled(); will(returnValue(true));
+                oneOf(logger).isTraceEnabled(); will(returnValue(false));
+                oneOf(sessionEx).getService(); will(returnValue(service));
+                oneOf(sessionEx).getRemoteAddress(); will(returnValue(remoteAddress));
+                oneOf(sessionEx).getSubject(); will(returnValue(subject));
+                oneOf(sessionEx).getLocalAddress(); will(returnValue(localAddress));
+                oneOf(sessionEx).getFilterChain(); will(returnValue(filterChain));
+                oneOf(filterChain).addLast(with("http#logging"), with(any(ExceptionLoggingFilter.class)));
+                will(saveParameter(filter, 1));
+            }
+        });
+
+        assertTrue(LoggingFilter.addIfNeeded(logger, sessionEx, "http"));
+        assertEquals("http#%s 111.122.133.144:41234", filter.get().getFormat());
+    }
+
+    @Test
+    public void addIfNeeded_shouldAddObjectLoggingFilter() throws Exception {
+        final IoFilterChain filterChain = context.mock(IoFilterChain.class, "filterChain");
+        final IoAcceptorEx service = context.mock(IoAcceptorEx.class, "service");
+        ResourceAddressFactory addressFactory = ResourceAddressFactory.newResourceAddressFactory();
+        final ResourceAddress localAddress = addressFactory.newResourceAddress("http://111.122.133.144:2121/jms");
+        final ResourceAddress  remoteAddress = addressFactory.newResourceAddress("http://111.122.133.144:41234/jms");
+        final Subject subject = new Subject();
+
+        AtomicReference<LoggingFilter> filter = new AtomicReference<LoggingFilter>();
+
+        context.checking(new Expectations() {
+            {
+                oneOf(logger).isInfoEnabled(); will(returnValue(true));
+                oneOf(logger).isTraceEnabled(); will(returnValue(true));
+                oneOf(sessionEx).getService(); will(returnValue(service));
+                oneOf(sessionEx).getRemoteAddress(); will(returnValue(remoteAddress));
+                oneOf(sessionEx).getSubject(); will(returnValue(subject));
+                oneOf(sessionEx).getLocalAddress(); will(returnValue(localAddress));
+                oneOf(sessionEx).getFilterChain(); will(returnValue(filterChain));
+                oneOf(filterChain).addLast(with("http#logging"), with(any(ObjectLoggingFilter.class)));
+                will(saveParameter(filter, 1));
+            }
+        });
+
+        assertTrue(LoggingFilter.addIfNeeded(logger, sessionEx, "http"));
+        assertEquals("http#%s 111.122.133.144:41234", filter.get().getFormat());
+    }
+
+    @Test
     public void shouldLogExceptionWithStack() throws Exception {
         LoggingFilter filter = new ExceptionLoggingFilter(logger, "tcp%s");
         final Exception exception = new NullPointerException();
@@ -71,7 +130,7 @@ public class LoggingFilterTest {
             {
                 oneOf(logger).isInfoEnabled(); will(returnValue(true));
                 oneOf(session).getId(); will(returnValue(123L));
-                oneOf(logger).info(with(stringMatching(".*NullPointerException.*")), with(exception));
+                oneOf(logger).info(with(stringMatching("\\[tcp123].*NullPointerException.*")), with(exception));
                 oneOf(nextFilter).exceptionCaught(session, exception);
             }
         });
@@ -117,206 +176,11 @@ public class LoggingFilterTest {
     }
 
     @Test
-    public void getUserIdentifierShouldReturnLocalAddress() throws Exception {
-
-        final IoConnectorEx  connector = context.mock(IoConnectorEx .class, "connector");
-
-        final InetSocketAddress address = new InetSocketAddress(InetAddress.getByName("localhost"), 2121);
-
-        context.checking(new Expectations() {
-            {
-                oneOf(sessionEx).getService(); will(returnValue(connector));
-                oneOf(sessionEx).getLocalAddress(); will(returnValue(address));
-                oneOf(sessionEx).getRemoteAddress(); will(returnValue(address));
-            }
-        });
-        assertEquals("localhost:2121", LoggingFilter.getUserIdentifier(sessionEx));
-    }
-
-    @Test
-    public void getUserIdentifierShouldReturnRemoteIpv4AddressForIoAcceptor() throws Exception {
-
-        final IoAcceptorEx  service = context.mock(IoAcceptorEx .class, "service");
-
-        final InetSocketAddress address = new InetSocketAddress(InetAddress.getByAddress(new byte[]{127,0,0,1}), 2121);
-
-        context.checking(new Expectations() {
-            {
-                oneOf(sessionEx).getService(); will(returnValue(service));
-                oneOf(sessionEx).getRemoteAddress(); will(returnValue(address));
-                oneOf(sessionEx).getLocalAddress(); will(returnValue(address));
-            }
-        });
-        assertEquals("127.0.0.1:2121", LoggingFilter.getUserIdentifier(sessionEx));
-    }
-
-    @Test
-    public void getUserIdentifierShouldReturnIpv6AddressForIoAcceptor() throws Exception {
-
-        final IoAcceptorEx  service = context.mock(IoAcceptorEx .class, "service");
-
-        final InetSocketAddress address = new InetSocketAddress(InetAddress.getByAddress(
-                new byte[]{(byte) 0xfe, (byte) 0x80, 0, 0, 0, 0, 0, 0, (byte) 0x90,
-                        (byte) 0xea, 0x3e, (byte) 0xe4, 0x77, (byte) 0xad, 0x77, (byte) 0xec}), 2121);
-
-        context.checking(new Expectations() {
-            {
-
-                oneOf(sessionEx).getService(); will(returnValue(service));
-                oneOf(sessionEx).getRemoteAddress(); will(returnValue(address));
-                oneOf(sessionEx).getLocalAddress(); will(returnValue(address));
-            }
-        });
-        assertEquals("fe80:0:0:0:90ea:3ee4:77ad:77ec:2121", LoggingFilter.getUserIdentifier(sessionEx));
-    }
-
-    @Test
-    public void getUserIdentifierShouldReturnScopedIpv6AddressForIoAcceptor() throws Exception {
-
-        final IoAcceptorEx  service = context.mock(IoAcceptorEx .class, "service");
-
-        final InetSocketAddress address = new InetSocketAddress(Inet6Address.getByAddress(
-                null,
-                new byte[]{(byte) 0xfe, (byte) 0x80, 0, 0, 0, 0, 0, 0, (byte) 0x90, (byte) 0xea, 0x3e,
-                           (byte) 0xe4, 0x77, (byte) 0xad, 0x77, (byte) 0xec},
-                15),
-                2121);
-
-        context.checking(new Expectations() {
-            {
-                oneOf(sessionEx).getService(); will(returnValue(service));
-                oneOf(sessionEx).getRemoteAddress(); will(returnValue(address));
-                oneOf(sessionEx).getLocalAddress(); will(returnValue(address));
-            }
-        });
-        assertEquals("fe80:0:0:0:90ea:3ee4:77ad:77ec%15:2121", LoggingFilter.getUserIdentifier(sessionEx));
-    }
-
-    @Test
-    public void getUserIdentifierShouldReturnRemoteAddressForBridgeAcceptor() throws Exception {
-        context.setImposteriser(ClassImposteriser.INSTANCE);
-
-        final AbstractBridgeAcceptor<?, ?> service = context.mock(AbstractBridgeAcceptor.class, "service");
-
-        final InetSocketAddress address = new InetSocketAddress(InetAddress.getByAddress(new byte[]{127,0,0,1}), 2121);
-
-        context.checking(new Expectations() {
-            {
-                oneOf(sessionEx).getService(); will(returnValue(service));
-                oneOf(sessionEx).getRemoteAddress(); will(returnValue(address));
-                oneOf(sessionEx).getLocalAddress(); will(returnValue(address));
-            }
-        });
-        assertEquals("127.0.0.1:2121", LoggingFilter.getUserIdentifier(sessionEx));
-    }
-
-    @Test
-    public void getUserIdentifierShouldReturnTcpEndpointFromTransport() throws Exception {
-
-        ResourceAddressFactory addressFactory = ResourceAddressFactories.newResourceAddressFactory();
-        String transportURI = "tcp://localhost:2121";
-        final ResourceAddress transport = addressFactory.newResourceAddress(transportURI);
-
-        final IoAcceptorEx  service = context.mock(IoAcceptorEx .class, "service");
-        final Subject subject = new Subject();
-
-        context.checking(new Expectations() {
-            {
-                oneOf(sessionEx).getService(); will(returnValue(service));
-                oneOf(sessionEx).getLocalAddress(); will(returnValue(transport));
-                oneOf(sessionEx).getSubject(); will(returnValue(subject));
-                oneOf(sessionEx).getRemoteAddress(); will(returnValue(transport));
-            }
-        });
-        assertEquals("127.0.0.1:2121", LoggingFilter.getUserIdentifier(sessionEx));
-    }
-
-    @Test
-    public void getUserIdentifierShouldReturnTcpEndpointFromHttpAddress() throws Exception {
-        ResourceAddressFactory addressFactory = ResourceAddressFactories.newResourceAddressFactory();
-        String addressURI = "http://localhost:2121/jms";
-        final ResourceAddress address = addressFactory.newResourceAddress(addressURI);
-        final IoServiceEx service = context.mock(IoServiceEx.class, "service");
-        final Subject subject = new Subject();
-        context.checking(new Expectations() {
-            {
-                oneOf(sessionEx).getService(); will(returnValue(service));
-                oneOf(sessionEx).getLocalAddress(); will(returnValue(address));
-                oneOf(sessionEx).getSubject(); will(returnValue(subject));
-                oneOf(sessionEx).getRemoteAddress(); will(returnValue(address));
-            }
-        });
-
-        assertEquals("127.0.0.1:2121", LoggingFilter.getUserIdentifier(sessionEx));
-    }
-
-    @Test
-    public void getUserIdentifierShouldReturnPrincipalFromHttpAddressIdentifierSet() throws Exception {
-        ResourceAddressFactory addressFactory = ResourceAddressFactories.newResourceAddressFactory();
-        String addressURI = "http://localhost:2121/jms";
-        ResourceOptions options = ResourceOptions.FACTORY.newResourceOptions();
-        buildIdentityResolverOption(options);
-
-        final ResourceAddress address = addressFactory.newResourceAddress(addressURI, options);
-        final IoServiceEx service = context.mock(IoServiceEx.class, "service");
-        final Subject subject = buildSubject();
-        context.checking(new Expectations() {
-            {
-                oneOf(sessionEx).getService(); will(returnValue(service));
-                oneOf(sessionEx).getLocalAddress(); will(returnValue(address));
-                oneOf(sessionEx).getSubject(); will(returnValue(subject));
-                oneOf(sessionEx).getRemoteAddress(); will(returnValue(address));
-            }
-        });
-        assertEquals("test 127.0.0.1:2121", LoggingFilter.getUserIdentifier(sessionEx));
-    }
-
-    @Test
-    public void getUserIdentifierShouldReturnTcpEndpointFromWsAddress() throws Exception {
-        ResourceAddressFactory addressFactory = ResourceAddressFactories.newResourceAddressFactory();
-        String addressURI = "ws://localhost:2121/jms";
-        final SocketAddress address = addressFactory.newResourceAddress(addressURI);
-
-        final IoServiceEx service = context.mock(IoServiceEx.class, "service");
-        final Subject subject = new Subject();
-        context.checking(new Expectations() {
-            {
-                oneOf(sessionEx).getService(); will(returnValue(service));
-                oneOf(sessionEx).getLocalAddress(); will(returnValue(address));
-                oneOf(sessionEx).getSubject(); will(returnValue(subject));
-                oneOf(sessionEx).getRemoteAddress(); will(returnValue(address));
-            }
-        });
-        assertEquals("127.0.0.1:2121", LoggingFilter.getUserIdentifier(sessionEx));
-    }
-
-    @Test
-    public void getUserIdentifierShouldReturnPrincipalFromWsAddressIdentifierSet() throws Exception {
-        ResourceAddressFactory addressFactory = ResourceAddressFactories.newResourceAddressFactory();
-        String addressURI = "ws://localhost:2121/jms";
-        ResourceOptions options = ResourceOptions.FACTORY.newResourceOptions();
-        buildIdentityResolverOption(options);
-
-        final SocketAddress address = addressFactory.newResourceAddress(addressURI, options);
-        final IoServiceEx service = context.mock(IoServiceEx.class, "service");
-        final Subject subject = buildSubject();
-        context.checking(new Expectations() {
-            {
-                oneOf(sessionEx).getService(); will(returnValue(service));
-                oneOf(sessionEx).getLocalAddress(); will(returnValue(address));
-                oneOf(sessionEx).getSubject(); will(returnValue(subject));
-                oneOf(sessionEx).getRemoteAddress(); will(returnValue(address));
-            }
-        });
-        assertEquals("test 127.0.0.1:2121", LoggingFilter.getUserIdentifier(sessionEx));
-    }
-
-    @Test
     public void shouldAddLoggingFilterWhenUserIdIsScopedIpv6Address() throws Exception {
 
         final IoConnectorEx  connector = context.mock(IoConnectorEx .class, "connector");
         final IoFilterChain filterChain = context.mock(IoFilterChain.class, "filterChain");
-        
+
         final InetSocketAddress address = new InetSocketAddress(Inet6Address.getByAddress(
                 null,
                 new byte[]{(byte) 0xfe, (byte) 0x80, 0, 0, 0, 0, 0, 0, (byte) 0x90, (byte) 0xea, 0x3e,
@@ -337,30 +201,6 @@ public class LoggingFilterTest {
             }
         });
         LoggingFilter.addIfNeeded(logger, sessionEx, "tcp");
-    }
-
-    /**
-     * Method building identity resolver option
-     * @param options
-     */
-    private void buildIdentityResolverOption(ResourceOptions options) {
-        Collection<Class<? extends Principal>> realmUserPrincipalClasses = new ArrayList<>();
-        realmUserPrincipalClasses.add(DefaultUserConfig.class);
-        IdentityResolver httpIdentityResolver = new HttpIdentityResolver(realmUserPrincipalClasses );
-        options.setOption(IDENTITY_RESOLVER, httpIdentityResolver);
-    }
-
-    /**
-     * Method building subject set with default principal
-     * @return
-     */
-    private Subject buildSubject() {
-        final Subject subject = new Subject();
-        DefaultUserConfig defaultPrincipal = new DefaultUserConfig();
-        defaultPrincipal.setName("test");
-        defaultPrincipal.setPassword("test");
-        subject.getPrincipals().add(defaultPrincipal);
-        return subject;
     }
 
 }
