+++ util/pom.xml
@@ -33,6 +33,11 @@
             <artifactId>junit</artifactId>
         </dependency>
         <dependency>
+            <groupId>org.jmock</groupId>
+            <artifactId>jmock-junit4</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
             <groupId>org.kaazing</groupId>
             <artifactId>test.util</artifactId>
             <version>${project.version}</version>
